Refactor Auth & Webhook controllers into single ClerkController
